fixed tests for tox
